fix: circular dependency in workflow
