+++ packages/features/ee/workflows/lib/getAllWorkflows.ts
@@ -1,5 +1,9 @@
+import getOrgIdFromMemberOrTeamId from "@calcom/lib/getOrgIdFromMemberOrTeamId";
+import { getTeamIdFromEventType } from "@calcom/lib/getTeamIdFromEventType";
 import prisma from "@calcom/prisma";
-import type { WorkflowType } from "@calcom/prisma/client";
+import type { Prisma, WorkflowType } from "@calcom/prisma/client";
+import { WorkflowType as PrismaWorkflowType } from "@calcom/prisma/enums";
+import { EventTypeMetaDataSchema } from "@calcom/prisma/zod-utils";
 
 import type { Workflow } from "./types";
 
@@ -137,3 +141,51 @@ export const getAllWorkflows = async ({
 
   return workflows;
 };
+
+export async function getAllWorkflowsFromEventType(
+  eventType: {
+    workflows?: {
+      workflow: Workflow;
+    }[];
+    teamId?: number | null;
+    parentId?: number | null;
+    parent?: {
+      id?: number | null;
+      teamId: number | null;
+    } | null;
+    metadata?: Prisma.JsonValue;
+  } | null,
+  userId?: number | null
+) {
+  if (!eventType) return [];
+
+  const eventTypeWorkflows = eventType?.workflows?.map((workflowRel) => workflowRel.workflow) ?? [];
+
+  const teamId = await getTeamIdFromEventType({
+    eventType: {
+      team: { id: eventType?.teamId ?? null },
+      parentId: eventType?.parentId || eventType?.parent?.id || null,
+    },
+  });
+
+  const orgId = await getOrgIdFromMemberOrTeamId({ memberId: userId, teamId });
+
+  const isManagedEventType = !!eventType?.parent;
+
+  const eventTypeMetadata = EventTypeMetaDataSchema.parse(eventType?.metadata || {});
+
+  const workflowsLockedForUser = isManagedEventType
+    ? !eventTypeMetadata?.managedEventConfig?.unlockedFields?.workflows
+    : false;
+
+  const allWorkflows = await getAllWorkflows({
+    entityWorkflows: eventTypeWorkflows,
+    userId,
+    teamId,
+    orgId,
+    workflowsLockedForUser,
+    type: PrismaWorkflowType.EVENT_TYPE,
+  });
+
+  return allWorkflows;
+}
